Pull ticks out into their own ui-canvas
I suspect that I'll want a mini-map soon and even if I don't this
component could be refactored into a mini global context thing instead
It took me an embarrassingly long time to work out that my ticks stopped
when my ui-canvas disappeared not because my application hung, but because
that's where the ticks were being incremented. Doh!