Pulling out some of the ubergraph sepecific parts of graph
Technically graph should be generic graph stuff, ubergraph should
implement specific functionality to provide whatever functionality
is needed to allow graph to work and economy for example should just
have graph based game logic.

I'm not going to go there yet, because I'm not sure if I need all
of that, but I'd like to have some ubergraph stuff in graph.ubergraph ns
so that I can require it in graph.viz.

So that's the main reason for this. Hence there will still be ubergraph
stuff in graph, though as the project gets more complex, it should get more
generic.