Try adding RELEASE_VERSION as env and grab in script
jpackage error's on `v`'s in `--app-version`:
```
"Version [v0.0.1] contains invalid component [v0]"
```

So stripping it out.

Turns out the `#refs/*/` in `${GITHUB_REF#refs/*/}` means bash shell parameter prefix substitution.
So `#` or `##` followed by a pattern will remove that from the front of the `VAR ` and `%` or `%%` from the back.