add macos code sigining
Based on instructions from:
https://localazy.com/blog/how-to-automatically-sign-macos-apps-using-github-actions

In this case [though this](https://stackoverflow.com/questions/62142092/is-it-okay-to-use-github-secrets-with-a-public-repo) suggests that we should mask generated strings of passwords, namely `certificate.p12`, as it's a file that I'm not echoing out I think it's ok.

`$PASSWORD` is just a random password that I'm making for the keychain, so it should also be ok as it's not really a secret?

[Additional context regarding github masking and issues around it are here, hopefully that improves](https://github.com/actions/runner/issues/475#issuecomment-636238383)