Mocked up a write to one canvas, copy to another and then read test
We want to store our images on something, which looks like it will be a `Surface`, so we can reduce the amount of drawing required.

So this is testing that we can transfer the canvas from one to another.

By changing the 0 0 arguments on the end of `.drawImage` we can also implement a sprite system ^_^...