Noodling around how to simulate decisions
A lot of stuff that I see that simulates states etc seems to be derived from looking at stuff like civilisation and other "god" games, which is great in terms of the level of detail in the sim, as it's all sort of working at the level of the civ, but from the perspective of it being a simulation I kind of feel like it's not very interesting?
It lacks that added detail of who's making the choices and what do the biases and perspectives that they have bring to the choices they make?
Another way of thinking about it is looking at something like [Boatmurdered](https://lparchive.org/Dwarf-Fortress-Boatmurdered/Introduction/), that was a succession based game, so as this fortress was handed off from person to person, decisions were reversed, reinforced or there was just a change in direction.
Now that's interesting!
So we have space which we can offer players choices, this AI is a leader of this civ, or manages it's exploration or trading, is their successor more timid, or adventurous, or reckless? Even if we give you tools to change things, we need to give you the drive to want to change stuff!
So this is about trying to create a space where that can happen.