Created a notion of decisions and sub-decisions
Just a straight rewrite of the idea splitting things into high level *decisions* and sub-decisions called *choices*.

Need to do a cleanup of the print code.

There's already an issue with some sub-decisions not being possible, so the decisions should have been eliminated earlier, for example deciding to claim territory, except the only territory available to claim is ocean terrain, and the civ can't claim that. Basically should the decision have been made in the first place? From a this is a bug perspective, I was originally thinking that this should have been eliminated earlier in the chain, ie high level agents should remove invalid decisions from the pool of candidates and I still believe that is the case for now, but it raises an interesting thing, which is two possible additional behaviours, leaving that in means you can now have additional depth at both levels:
- Leaders from micromanagers who eliminate choices giving their reports only a small subset of the available options to careful leaders who just remove obviously bad options to thoughtless leaders who just say get it done, giving free reign to their subordinates.
- Space for the subordinates to simply make a choice from the options available to them, say none of the options are valid giving and just throwing their arms up saying nothing to be done here, because they don't like any of the options or they want to block progress for a turn, to subversively selecting a valid option which was eliminated earlier by their boss.