Add `:nonviable` as a decision
So thinking about this, I should probably move so that each tick is a month, that means that even though the person making the decision and the person making the sub-decision is the same person, we can have a situation where the decision is for next month and the sub-decision is last month's decision. Where the default action can be `:planning` in the case of game start and when a decision is `:nonviable`.