Just threw together a simple implementation for the decisions
I've not written any of the `log-history` entries for the moment, I'm sort of tempted to allow `log-history` to capture additional data which it can use to do more than show a user log message. Though perhaps that's a bit of over-engineering. I won't do it for now, but I'll come back to write up the log messages later.