Created **mock** `leader-tick` and `subordinate-tick` + `Next Steps` in intro.md
An update to 57d19be, wrote a **mock** `leader-tick` and `subordinate-tick` function, primarily to sketch out how they should behave, we're being explicit about the leader making a decision and the subordinate making a choice and executing on the decision. This might be split further into `leader`, `subordinate`, `worker` for example, as roles, but that feels a bit too much like planning for the future.

Added a segment called `Next Steps` in `docs/intro.md` so that we can put together something of small todo list.

Not a complex plan atm:
1. Figure out how to make `leader-tick` and `subordinate-tick` actually run in the game so the world changes.
2. Denormalise bits to clean up some of the deeper nesting, for example pull `peeps` out to the top level as well as `civs`.
3. Spend a bit of time thinking about where we might draw the system boundaries, if this game offered "services", what would those be?