Start sketching out a db layer
The intention here is not to write a whole db layer, that would be silly, the point is to write the minimal amount to work out what a db api would look like and start hooking it in.

We can then pull a more mature db and write an adaptor to that, this way we minimise any effect of the db concerns from overly impacting the game datastructures.