Start trying to write a test case
So we've got an obvious first step and we've kind of got an idea of the very simple way this magical api that we'd like to work within our system should work at a really high level.
So now the next step is to actually make this api work somehow.