Added `:civ/territory` to begin thinking of replacing `::civ/territory` and `::civ/territory->development`
This is getting really bad, I know ns qualified keywords are good practice, but I'm probably not going to qualify them with the full project name for the moment. I'll shorten them, it's getting pretty unreadable in a lot of places.

Hmm, though thinking about it maybe it's better to create a sort of data namespace that I can safely import everywhere which can just hold names. Need to figure out a good name for it...

Other than that the big issue is needing to tidy up the datascript db, there's still a lot of messy stuff in there which I really should clean up cruft-wise, lots of datastructures that existed as indexes for example which are for the moment better handled by datascript itself...
So this is starting that a little.