Experiment with caching render
Added a place to store surface, canvas and image data against a keyword, so the idea is to specify a name, and then draw that as a layer only once every few frames and then render that the rest of the frames.

It would be better if this was *smart* about when stuff got dirty etc, but for now let's not worry about that.

This might also be a good starting point to test creating a sort of asset library, rendering text seems to be much slower than rendering an image rect, so theoretically I should also capture images of text segments and just render them again.

Looking at how HumbleUI does `dynamic` is instructive here.
Not sure if it would be better to try using a dynamic-like thing for each cell? I should benchmark it.