Add `rewrite` and `infer` from minikusari
From: https://github.com/frankiesardo/minikusari

Thanks to @frankiesardo for creating this!

It's been inlined because it's not available currently on clojars and due to the [Apache License 2.0](https://github.com/frankiesardo/minikusari/blob/main/LICENSE) I should be able to use it with a reference to it.

Not sure how long I'll use it for, but it's useful for now!