Change names `:loc` -> `:coord` `:pos` -> `:place`
`:loc` -> `:coord`:  This way `:loc` can stand for location, as in the location the entity is at, while a 2d point can be a coordinate.

`:pos` -> `:place`: Once again, refer to it as place instead of  `:pos` (position), hopefully implying that place isn't just a point, but can contain more attributes.