Tax by pulling the money out and paying later
As a reminder, the original comment was:
```
;; TODO: Is this a good idea?
It may be nice to have players earn on a period, but the sim's don't know to save enough money for tax.
Which means either modelling debt, or taking money out of the system and only paying the tax as the end.
```

So we're now collecting money and giving it out at the end of the month.