Make `:close-price` not equal `:current-price`
If the number of trades span a wide range, then it felt like this can cause problems, so instead we're taking the average of the prices, weighted by their size, across the entire accepted trades as the current price.

In markets with low volatility, ie only one trade now and again, this is the same behaviour as before.